Renamed modules and updated imports, reran notebooks
