+++ notebooks/01_basic_usage.py
@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.14.1
+#       jupytext_version: 1.14.5
 #   kernelspec:
 #     display_name: Python 3.9.4 ('rdkit')
 #     language: python
@@ -42,7 +42,7 @@ mols = [Chem.MolFromSmiles(smiles) for smiles in smiles_strings]
 # Next we import the Morgan fingerprint transformer
 
 # %%
-from scikit_mol.transformers import MorganTransformer
+from scikit_mol.fingerprints import MorganTransformer
 
 transformer = MorganTransformer(radius=3)
 print(transformer)
@@ -91,7 +91,7 @@ transformer.transform(mols, y)
 # Sometimes we may want to transform SMILES into molecules, and scikit-mol also has a transformer for that. It simply takes a list of SMILES and produces a list of RDKit molecules, this may come in handy when building pipelines for machine learning models, as we will demo in another notebook.
 
 # %%
-from scikit_mol.transformers import SmilesToMol
+from scikit_mol.conversions import SmilesToMol
 smi2mol = SmilesToMol()
 print(smi2mol)
